requirements.txt
typer[all]==0.9.0
pyyaml==6.0.1
tiktoken==0.5.2
rich==13.7.0
